feat: 스토어 평균가격 추가로 인한 로직 수정, menu price 필드 int로 변경, 스토어 검색 조건 추가

+++ fooding-api/src/main/java/im/fooding/app/dto/response/user/menu/MenuResponse.java
@@ -61,7 +61,7 @@ public class MenuResponse {
                 .name(menu.getName())
                 .description(menu.getDescription())
                 .imageUrl(menu.getImageUrl())
-                .price(menu.getPrice().intValue())
+                .price(menu.getPrice())
                 .sortOrder(menu.getSortOrder())
                 .isSignature(menu.isSignature())
                 .isRecommend(menu.isRecommend())
